chore: moved location-adornment to location directory
