+++ frontend/src/features/location/components/location-adornment.tsx
