+++ frontend/src/features/location/components/location-search-bar.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+import React from 'react';
+import {SearchBar, SearchBarProps} from "@/features/search";
+
+type Props = SearchBarProps & {
+
+}
+
+export function LocationSearchBar ({ ...searchBarProps }: Props) {
+  return (
+    <>
+      <SearchBar {...searchBarProps} />
+    </>
+  );
+};
