+++ frontend/src/features/location/hooks/use-geolocation.ts
@@ -1,3 +1,5 @@
+"use client"
+
 import {useCallback, useState} from "react";
 import {Location} from "@/features/vendor";
