+++ frontend/src/features/location/index.ts
@@ -1,2 +1,6 @@
-export * from "./api/geolocation"
 export * from "./hooks/use-geolocation"
+export * from "./api/fetch-nearby-places"
+export * from "./hooks/use-nearby-vendors"
+
+export * from "./components/location-search-bar"
+export * from "./components/location-adornment"
