+++ frontend/src/features/search/index.ts
@@ -1,2 +1,1 @@
-export * from "./components/location-adornment"
 export * from "./components/search-bar"
